[REF] test_orm: remove base models dependencies from test_one2many
